+++ Sources/AutoSDK/AutoSDKVersion.m
@@ -1,4 +1,4 @@
 #import "include/AutoSDK.h"
 
-double AutoSDKVersionNumber = 1.30;
-const unsigned char AutoSDKVersionString[] = "1.30.2";
+double AutoSDKVersionNumber = 1.31;
+const unsigned char AutoSDKVersionString[] = "1.31.0";